app/requirements.txt
pandas == 1.4.2
Pillow == 9.1.1
fastapi == 0.78.0
pydantic == 1.9.1
numpy == 1.22.4
uvicorn == 0.17.6
boto3 == 1.24.9
python-multipart == 0.0.5
flask == 2.2.2